+++ Changelog.md
@@ -26,6 +26,7 @@ Rust:
 - in FBS, replace WireList by [WireRange] and remove type_id from WireRange
 - add conversion declarations in Relation
 - split public/private inputs files into one file per type
+- update conversion to be compliant with the last version of the specs
 
 # SIEVE IR v1.0.1, 2022-08